FIX: cut down code into function & del meat.py for main.py
